fix(domain): FileIdentityのlinux側で不要になる整数変換を落とす
PR CI の static ジョブ(ubuntu-24.04)で `make lint` が
`internal/domain/descriptor.go:130: unnecessary conversion (unconvert)`
で失敗していた。`syscall.Stat_t.Dev` は darwin では int32、linux では
uint64 なので、`uint64(stat.Dev)` は darwin では必要な変換だが linux
では自明に不要で、linux 解析だけが落ちる。ローカルの `make lint` は
darwin で走るため気付けていなかった(`CGO_ENABLED=0 GOOS=linux
golangci-lint run ./...` で再現する)。

`%d` はどちらの整数幅もそのまま書式化できるので、変換自体を落とす。
nolint やプラットフォーム別ファイルを増やすより小さい。この identity は
ローカルの inode を指すだけで、同一バイナリが取得した別の identity と
しか比較しないため、darwin と linux で表現が一致する必要はない。

確認: `golangci-lint run ./...` は darwin / GOOS=linux の両方で
0 issues、`deadcode -test ./...` も両方 0 件。static ジョブで lint より
後段にあり未実行だった mod-tidy-check / docs-check / workflow-check /
shell-check / fmt-check も個別に通した。

+++ internal/domain/descriptor.go
@@ -115,7 +115,12 @@ func OpenRootAt(owner *os.Root, relative string) (*os.Root, error) {
 // package builds only for darwin and linux (see physical_unix.go), so the
 // concrete type behind info.Sys() is always *syscall.Stat_t; a direct type
 // assertion replaces the reflection this used to need only to smooth over
-// Dev/Ino's differing integer widths between those two platforms. A failed
+// Dev/Ino's differing integer widths between those two platforms; %d formats
+// either width directly, so no explicit conversion is involved (Dev is int32
+// on darwin and uint64 on linux, and converting would be flagged as
+// unnecessary on linux). The identity only ever names a local inode and is
+// only compared against another identity captured by the same binary, so the
+// per-platform rendering does not have to agree across platforms. A failed
 // assertion is fail-closed: it returns an error rather than a zero identity,
 // since a zero identity would compare equal to another failure instead of
 // being treated as unavailable.
@@ -127,5 +132,5 @@ func FileIdentity(info os.FileInfo) (string, error) {
 	if !ok || stat == nil {
 		return "", fmt.Errorf("unsupported file identity type %T", info.Sys())
 	}
-	return fmt.Sprintf("%d:%d", uint64(stat.Dev), stat.Ino), nil
+	return fmt.Sprintf("%d:%d", stat.Dev, stat.Ino), nil
 }
